+++ libft/ft_strlcpy.c
@@ -6,14 +6,15 @@
 /*   By: mcolomba <mcolomba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/23 14:19:44 by mcolomba          #+#    #+#             */
-/*   Updated: 2022/09/27 14:10:11 by mcolomba         ###   ########.fr       */
+/*   Updated: 2022/11/03 11:11:47 by mcolomba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, const char *src,
-					unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	srclen;
+	size_t	srclen;
 
 	srclen = ft_strlen(src);
 	if (srclen + 1 < size)
